moving accessor stuff around
