+++ src/wmtk/attribute/Accessor.hpp
@@ -2,7 +2,11 @@
 
 #include "TupleAccessor.hpp"
 
+#include <Eigen/Dense>
+
 namespace wmtk::attribute {
+
 template <typename T>
-using ConstAccessor = const TupleAccessor<T>;
-} // namespace wmtk::attribute
+class Accessor: public TupleAccessor<T> {};
+
+} // namespace wmtk
