+++ src/wmtk/multimesh/utils/tuple_map_attribute_io.hpp
@@ -1,4 +1,4 @@
-#include <wmtk/Accessor.hpp>
+#include <wmtk/attribute/Accessor.hpp>
 #include <wmtk/Types.hpp>
 
 // #define WMTK_DISABLE_COMPRESSED_MULTIMESH_TUPLE
@@ -24,18 +24,18 @@ TwoTupleVector tuples_to_vectors(const Tuple& a, const Tuple& b);
 
 // utility functions to simplify how we encode 2-Tuple attributes as 10-int64_t attribute
 void write_tuple_map_attribute(
-    Accessor<int64_t>& map_accessor,
+    wmtk::attribute::Accessor<int64_t>& map_accessor,
     const Tuple& source_tuple,
     const Tuple& target_tuple);
 
 std::tuple<Tuple, Tuple> read_tuple_map_attribute(
-    const ConstAccessor<int64_t>& accessor,
+    const wmtk::attribute::Accessor<int64_t>& accessor,
     const Tuple& source_tuple);
 
 
 void symmetric_write_tuple_map_attributes(
-    Accessor<int64_t>& a_to_b,
-    Accessor<int64_t>& b_to_a,
+    wmtk::attribute::Accessor<int64_t>& a_to_b,
+    wmtk::attribute::Accessor<int64_t>& b_to_a,
     const Tuple& a_tuple,
     const Tuple& b_tuple);
 void write_tuple_map_attribute_slow(
