+++ src/wmtk/operations/utils/UpdateEdgeOperationMultiMeshMapFunctor.cpp
@@ -887,20 +887,20 @@ void UpdateEdgeOperationMultiMeshMapFunctor::operator()(
 }
 
 int64_t UpdateEdgeOperationMultiMeshMapFunctor::child_global_cid(
-    const attribute::ConstAccessor<int64_t>& parent_to_child,
+    const attribute::Accessor<int64_t>& parent_to_child,
     int64_t parent_gid) const
 {
     return MultiMeshManager::child_global_cid(parent_to_child, parent_gid);
 }
 int64_t UpdateEdgeOperationMultiMeshMapFunctor::parent_global_cid(
-    const attribute::ConstAccessor<int64_t>& child_to_parent,
+    const attribute::Accessor<int64_t>& child_to_parent,
     int64_t child_gid) const
 {
     return MultiMeshManager::parent_global_cid(child_to_parent, child_gid);
 }
 
 int64_t UpdateEdgeOperationMultiMeshMapFunctor::parent_local_fid(
-    const attribute::ConstAccessor<int64_t>& child_to_parent,
+    const attribute::Accessor<int64_t>& child_to_parent,
     int64_t child_gid) const
 {
     return MultiMeshManager::parent_local_fid(child_to_parent, child_gid);
