+++ src/wmtk/operations/utils/UpdateVertexMultiMeshMapHash.cpp
@@ -9,7 +9,7 @@
 #include <wmtk/simplex/closed_star.hpp>
 
 namespace wmtk::operations::utils {
-void update_vertex_operation_hashes(Mesh& m, const Tuple& vertex, Accessor<int64_t>& hash_accessor)
+void update_vertex_operation_hashes(Mesh& m, const Tuple& vertex, attribute::Accessor<int64_t>& hash_accessor)
 {
     const PrimitiveType pt = m.top_simplex_type();
     const simplex::SimplexCollection star =
@@ -65,7 +65,7 @@ void update_vertex_operation_hashes(Mesh& m, const Tuple& vertex, Accessor<int64
 void update_vertex_operation_multimesh_map_hash(
     Mesh& m,
     const simplex::SimplexCollection& vertex_closed_star,
-    Accessor<int64_t>& parent_hash_accessor)
+    attribute::Accessor<int64_t>& parent_hash_accessor)
 {
     auto& mm_manager = m.m_multi_mesh_manager;
